requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
